+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and this file attempts to follow the guidelines provided by [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased]
-TODO.
+### Added
+- Add extension to directly dequeue a reusable cell via a `TableViewCell` instance.
 
 ## [1.0.0] - 2016-03-21
 ### Changed